third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-reconnect-expected.txt

[Worker] Test EventSource reconnect after end of event stream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-reconnect.js
PASS [Worker] es.readyState is es.CONNECTING
PASS [Worker] es.readyState is es.OPEN
PASS [Worker] evt.lastEventId is "77☃"
PASS [Worker] es.readyState is es.CONNECTING
PASS [Worker] es.readyState is es.OPEN
PASS [Worker] evt.data is "77☃"
PASS [Worker] evt.lastEventId is "77☃"
PASS [Worker] es.readyState is es.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
